+++ pysat/instruments/pysat_testmodel.py
@@ -43,6 +43,13 @@ def init(self):
     return
 
 
+def clean(self):
+    """Cleaning function
+    """
+
+    pass
+
+
 def load(fnames, tag=None, inst_id=None, num_samples=None):
     """ Loads the test files